requirements.txt
pyyaml
rich
prometheus_client
